fixing details, adding swagger ui, replacing user model
