--- src/middlewares/authentificate.js
@@ -1,46 +1,0 @@
-
-import createHttpError from 'http-errors';
-
-import { SessionsCollection } from '../db/models/session.js';
-import { UsersCollection } from '../db/models/user.js';
-import { response } from 'express';
-
-export const authentificate = async (req, res, next) => {
-    const authHeader = req.get('Authorization');
-
-    if (!authHeader) {
-        next(createHttpError(401, 'Authorization token is required'));
-        return;
-    }
-
-    const bearer = authHeader.split(' ')[0];
-    const token = authHeader.split(' ')[1];
-
-    if (bearer !== 'Bearer' || !token) {
-        next(createHttpError(401, 'Authorization header must use the Bearer scheme'));
-        return;
-    }
-
-    const session = await SessionsCollection.findOne({ accessToken: token });
-
-    if (!session) {
-        next(createHttpError(401, 'Session not found'));
-        return;
-    }
-
-    const isAccessTokenExpired = new Date() > new Date(session.accessTokenValidUntil);
-
-    if (isAccessTokenExpired) {
-        next(createHttpError(401, 'Access token expired'));
-    }
-
-    const user = await UsersCollection.findById(session.userId);
-
-    if (!user) {
-        next(createHttpError(401));
-        return;
-    }
-
-    req.user = user;
-    next();
-};
