+++ src/validation/auth.js
@@ -1,9 +1,21 @@
 import Joi from 'joi';
+import dayjs from 'dayjs';
+
+// поточна дата
+const curDate = dayjs();
+
+// мінімальна дата – через 1 тиждень
+const minDate = curDate.add(1, 'week').toDate();
+
+// максимальна дата – через 40 тижнів
+const maxDate = curDate.add(40, 'week').toDate();
 
 export const registerUserSchema = Joi.object({
     name: Joi.string().min(1).max(32).required(),
     email: Joi.string().email().max(64).required(),
     password: Joi.string().min(8).max(128).required(),
+    gender: Joi.string().valid('boy', 'girl').optional().allow(null),
+   dueDate: Joi.date().min(minDate).max(maxDate).optional(),
 });
 
 export const loginUserSchema = Joi.object({
