+++ swagger/components/responses/409.yaml
@@ -1,0 +1,15 @@
+description: Conflict - user already exists
+content:
+  application/json:
+    schema:
+      type: object
+      required:
+        - status
+        - message
+      properties:
+        status:
+          type: integer
+          example: 409
+        message:
+          type: string
+          example: User with this email already exists
