+++ swagger/paths/auth/login/post.yaml
@@ -1,0 +1,47 @@
+tags:
+  - Auth
+summary: Login user
+operationId: loginUser
+description: "Login user with email and password to get JWT token."
+requestBody:
+  required: true
+  content:
+    application/json:
+      schema:
+        type: object
+        required:
+          - email
+          - password
+        properties:
+          email:
+            type: string
+            format: email
+            example: "johndoe@example.com"
+          password:
+            type: string
+            example: "MySecret123"
+responses:
+  "200":
+    description: Login successful
+    content:
+      application/json:
+        schema:
+          type: object
+          required:
+            - status
+            - message
+            - token
+            - refreshToken
+          properties:
+            status:
+              type: integer
+              example: 200
+            message:
+              type: string
+              example: "Login successful"
+            token:
+              type: string
+              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
+            refreshToken:
+              type: string
+              example: "dGhpcy1pcy1hLXJlZnJlc2gtdG9rZW4..."
