+++ swagger/paths/auth/logout/post.yaml
@@ -1,0 +1,21 @@
+tags:
+  - Auth
+summary: Logout user
+operationId: logoutUser
+description: "Invalidate the current session (requires Authorization header)."
+security:
+  - bearerAuth: []
+responses:
+  "200":
+    description: Logout successful
+    content:
+      application/json:
+        schema:
+          type: object
+          properties:
+            status:
+              type: integer
+              example: 200
+            message:
+              type: string
+              example: "Successfully logged out!"
