+++ swagger/paths/auth/refresh/post.yaml
@@ -1,0 +1,24 @@
+tags:
+  - Auth
+summary: Refresh access token
+operationId: refreshUserSession
+description: "Refresh access token using refresh token (requires Authorization header)."
+security:
+  - bearerAuth: []
+responses:
+  "200":
+    description: Session refreshed successfully
+    content:
+      application/json:
+        schema:
+          type: object
+          properties:
+            status:
+              type: integer
+              example: 200
+            message:
+              type: string
+              example: "Successfully refreshed a session!"
+            token:
+              type: string
+              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
